requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
Flask==2.2.2
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
flask-talisman==1.1.0
Flask-Testing==0.8.1
Flask-WTF==1.2.1
importlib_resources==6.4.5
itsdangerous==2.1.2
jiter==0.5.0
limits==3.13.0
openai==1.45.1
ordered-set==4.1.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==2.2.2
wrapt==1.16.0
WTForms==3.1.2
